email-rag/requirements.txt
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pgvector>=0.3,<1.0
python-dotenv>=1.0,<2.0
click>=8.0,<9.0
tqdm>=4.60,<5.0
